refactor: 내부통신 컨트롤러 병합
